Fix build issue with linaro Arm64 compiler
With the Arm64 compiler downloaded from Linaro, I get a compile failure
due to a missing header file.  The file in question is in the same
directory as the uncompileable C file, so it appears to be an issue with
the default include path.  Fix this by explicitly including the
appropriate directory in the include path.