WIP: Sound: core driver
[Problem]
Currently hw_ptr will be increased by a buffer size if
1. expected hw ptr is larger then HW reported, and
2. time from last recorded time is over half a buffer time.
However, when HW reported pointer is only a few bytes smaller than
expected hw ptr, and period size is half of the buffer size,
this condition will be wrongly entered.

[Solution]
The original condition #2 should be wrong since in the end,
a buffer size is added to the hw ptr, not half a buffer size.
Only when the time difference is really over a buffer time then
one should add a buffer size to the hw ptr.

Change-Id: I4b573a40eb8432ecc3d7db7e082b8e226e0295b7
Signed-off-by: Koro Chen <koro.chen@mediatek.com>