FIXUP: ASoC: mediatek: add missing clocks
[Problem]
There are some clocks required for audio but not ultilized in CCF yet.
Before the new CCF version is available, we can only operate these clocks
directly by audio driver for now.

[Solution]
Add clock related registers, and set these registers directly. A debugfs
is also added to check the correctness of register settings.
Once the CCF is updated with the missing clocks, this FIXUP can be removed.

Change-Id: I540cefa51722087de269ecbaa88b13f8cd84aaee
Signed-off-by: Koro Chen <koro.chen@mediatek.com>